+++ ecms_base/modules/custom/ecms_blocks/src/Form/SearchStateBlockForm.php
@@ -1,0 +1,68 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace Drupal\ecms_blocks\Form;
+
+use Drupal\Core\Form\FormBase;
+use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\Url;
+use Drupal\Core\Routing\TrustedRedirectResponse;
+
+/**
+ * Provides a custom site search form.
+ *
+ * @package Drupal\ecms_blocks\Form
+ */
+class SearchStateBlockForm extends FormBase {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getFormId(): string {
+    return 'ecms_search_state_block_form';
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function buildForm(array $form, FormStateInterface $form_state): array {
+
+    $form['search_input'] = [
+      '#type' => 'textfield',
+      "#placeholder" => $this->t('Search RI Government'),
+      '#required' => TRUE,
+      "#attributes" => [
+        'type' => "search",
+      ],
+      '#theme_wrappers' => [],
+    ];
+
+    $form['actions']['#type'] = 'actions';
+
+    $form['actions']['submit'] = [
+      '#type' => 'submit',
+      '#value' => $this->t('Search'),
+    ];
+
+    return $form;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function submitForm(array &$form, FormStateInterface $form_state): void {
+    $input = $form_state->getValue('search_input');
+    $params['query'] = [
+      'q' => $input,
+    ];
+
+    $response = new TrustedRedirectResponse(Url::fromUri("https://www.ri.gov/search/result.php", $params)->toString());
+
+    $metadata = $response->getCacheableMetadata();
+    $metadata->setCacheMaxAge(0);
+
+    $form_state->setResponse($response);
+  }
+
+}